% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gihub.traffic.R
\name{fetch_github_traffic}
\alias{fetch_github_traffic}
\title{Fetch github traffic info from github}
\usage{
fetch_github_traffic(repos)
}
\arguments{
\item{repos}{A characer vector of repos in he "owner/repo" or "owner" format,
in the latter case the info of all repos is fetched.}
}
\value{
A list of ables
}
\description{
This provides a list of tables containing all the info github makes
available regarding traffic to a repo. It is limited to the last 14 days. To
store the data and update it, use \code{update_github_traffic_db()}, to automate it
(so your DB is updated daily for instance) use \code{schedule_github_traffic_db_updates()}
(Windows only). Then retrieve local DB with \code{github_traffic_db()}.
}
\details{
The output contains tibbles named :
\itemize{
\item clones_per_day
\item clones_per_week,
\item clones_per_14days
\item page_views_per_day
\item page_views_per_week,
\item page_views_per_14days
\item popular_paths
\item popular_referrers
\item repo_info.
}

Weeks start on monday.
}
