% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.traffic-package.R
\docType{package}
\name{github.traffic-package}
\alias{github.traffic}
\alias{github.traffic-package}
\title{github.traffic: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Antoine Fabri \email{antoine.fabri@gmail.com}

}
\keyword{internal}
