% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gihub.traffic.R
\name{github_traffic_db}
\alias{github_traffic_db}
\title{Github Traffic DB}
\usage{
github_traffic_db()
}
\description{
Returns the database as a list of tibbles.
}
\details{
The tibbles are named :
\itemize{
\item clones_per_day
\item clones_per_week,
\item clones_per_14days
\item page_views_per_day
\item page_views_per_week,
\item page_views_per_14days
\item popular_paths
\item popular_referrers
\item repo_info.
}
}
