% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gihub.traffic.R
\name{github_traffic_plot}
\alias{github_traffic_plot}
\title{Plot github traffic}
\usage{
github_traffic_plot(
  repos,
  what = c("clones", "views"),
  freq = c("day", "week")
)
}
\arguments{
\item{repos}{A vector of repos or owners}

\item{what}{"clones" or "views"}

\item{freq}{"day" or "week"}
}
\description{
If several repos or an owner is provided in the \code{repos} argument, \code{count}
and \code{uniques} will be summed, which might lead to a (probably small) overestimation
of \code{uniques}.
}
