% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gihub.traffic.R
\name{schedule_github_traffic_db_updates}
\alias{schedule_github_traffic_db_updates}
\title{Schedule Github Traffic DB Updates}
\usage{
schedule_github_traffic_db_updates(
  repos,
  schedule = "DAILY",
  taskname = "Update-Github-Traffic-DB",
  ...
)
}
\arguments{
\item{repos}{A characer vector of repos in he "owner/repo" or "owner" format,
in the latter case the info of all repos is fetched.}

\item{schedule}{When to schedule the update. Either one of 'ONCE', 'MONTHLY', 'WEEKLY', 'DAILY', 'HOURLY', 'MINUTE', 'ONLOGON', 'ONIDLE'}

\item{taskname}{Name of the task.Should not contain any spaces.}

\item{...}{additional parameers forwarded to \code{taskscheduleR::taskscheduler_create()}}
}
\value{
Returns \code{NULL} silently, called for side effects.
}
\description{
This is only for windows for now. Allows you to setup automatic updates
at the chosen frequency into the DB defined in the \code{"R_GITHUB_TRAFFIC_DB_PATH"},
or \code{"~/.github_traffic_db.RDS"} if it's undefined.
}
