% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gihub.traffic.R
\name{update_github_traffic_db}
\alias{update_github_traffic_db}
\title{Update Github Traffic DB}
\usage{
update_github_traffic_db(repos, append = TRUE)
}
\arguments{
\item{repos}{A characer vector of repos in he "owner/repo" or "owner" format,
in the latter case the info of all repos is fetched.}

\item{append}{whether to keep existing DB, \code{append = FALSE} will remove everyhing you've
saved so far!}
}
\description{
This fetches the latest info from github and adds it to your local DB.
You can then import your DB with : \code{github_traffic_db()}.
}
